+++ packages/ui-scaffold/src/components/announcement/index.ts
@@ -1,6 +1,0 @@
-export { AnnouncementUI } from "./announcement.ui";
-export { useAnnouncementScript } from "./announcement.script";
-export {
-  AnnouncementWidget,
-  type AnnouncementWidgetProps,
-} from "./announcement.widget";