+++ packages/ui-scaffold/src/components/main/mainNav.ui.tsx
@@ -5,6 +5,7 @@ import { WalletConnectButtonExtension } from "../accountMenu/menu.widget";
 import { AccountSummaryWidget } from "../accountSummary";
 import { ChainMenuWidget } from "../chainMenu";
 import { LanguageSwitcherWidget } from "../languageSwitcher";
+import { MessageCenterWidget } from "../messageCenter/msgCenter.widget";
 import { SubAccountWidget } from "../subAccount";
 import { CampaignButton } from "./campaignButton";
 import { LinkDeviceWidget } from "./linkDevice";
@@ -54,6 +55,7 @@ export const MainNav: FC<PropsWithChildren<MainNavScriptReturn>> = (props) => {
     const languageSwitcher = <LanguageSwitcherWidget />;
     const subAccount = showSubAccount && <SubAccountWidget />;
     const chainMenu = showChainMenu && <ChainMenuWidget />;
+    const notify = <MessageCenterWidget />;
     const walletConnect = !hideWalletConnectButton && (
       <WalletConnectButtonExtension />
     );
@@ -83,6 +85,7 @@ export const MainNav: FC<PropsWithChildren<MainNavScriptReturn>> = (props) => {
         languageSwitcher,
         subAccount,
         chainMenu,
+        notify,
         walletConnect,
       });
     }
@@ -119,6 +122,7 @@ export const MainNav: FC<PropsWithChildren<MainNavScriptReturn>> = (props) => {
               {linkDevice}
             </>
           )}
+          {notify}
           {languageSwitcher}
           {subAccount}
           {chainMenu}