+++ packages/ui-scaffold/src/components/messageCenter/msgCenter.widget.tsx
@@ -1,0 +1,14 @@
+import { useScaffoldContext } from "../scaffold/scaffoldContext";
+import { MessageCenter } from "./msgCenter.ui";
+
+export const MessageCenterWidget = () => {
+  const { announcementState } = useScaffoldContext();
+
+  return (
+    <MessageCenter
+      messages={announcementState.tips}
+      maintenanceDialogInfo={announcementState.maintenanceDialogInfo}
+      showAnnouncement={announcementState.showAnnouncement}
+    />
+  );
+};