+++ packages/default-solana-adapter/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @orderly.network/default-solana-adapter
 
+## 2.8.1
+
+### Patch Changes
+
+- publish
+- 0474e5f: publish
+- Updated dependencies
+- Updated dependencies [0474e5f]
+  - @orderly.network/core@2.8.1
+  - @orderly.network/types@2.8.1
+
+## 2.8.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.8.1-alpha.0
+  - @orderly.network/types@2.8.1-alpha.0
+
 ## 2.8.0
 
 ### Minor Changes