+++ packages/trading/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @orderly.network/trading
 
+## 2.8.1
+
+### Patch Changes
+
+- publish
+- 0474e5f: publish
+- Updated dependencies
+- Updated dependencies [0474e5f]
+  - @orderly.network/react-app@2.8.1
+  - @orderly.network/hooks@2.8.1
+  - @orderly.network/i18n@2.8.1
+  - @orderly.network/markets@2.8.1
+  - @orderly.network/portfolio@2.8.1
+  - @orderly.network/types@2.8.1
+  - @orderly.network/ui@2.8.1
+  - @orderly.network/ui-chain-selector@2.8.1
+  - @orderly.network/ui-connector@2.8.1
+  - @orderly.network/ui-leverage@2.8.1
+  - @orderly.network/ui-order-entry@2.8.1
+  - @orderly.network/ui-orders@2.8.1
+  - @orderly.network/ui-positions@2.8.1
+  - @orderly.network/ui-scaffold@2.8.1
+  - @orderly.network/ui-share@2.8.1
+  - @orderly.network/ui-tradingview@2.8.1
+  - @orderly.network/ui-transfer@2.8.1
+  - @orderly.network/utils@2.8.1
+
+## 2.8.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.8.1-alpha.0
+  - @orderly.network/hooks@2.8.1-alpha.0
+  - @orderly.network/i18n@2.8.1-alpha.0
+  - @orderly.network/markets@2.8.1-alpha.0
+  - @orderly.network/portfolio@2.8.1-alpha.0
+  - @orderly.network/types@2.8.1-alpha.0
+  - @orderly.network/ui@2.8.1-alpha.0
+  - @orderly.network/ui-chain-selector@2.8.1-alpha.0
+  - @orderly.network/ui-connector@2.8.1-alpha.0
+  - @orderly.network/ui-leverage@2.8.1-alpha.0
+  - @orderly.network/ui-order-entry@2.8.1-alpha.0
+  - @orderly.network/ui-orders@2.8.1-alpha.0
+  - @orderly.network/ui-positions@2.8.1-alpha.0
+  - @orderly.network/ui-scaffold@2.8.1-alpha.0
+  - @orderly.network/ui-share@2.8.1-alpha.0
+  - @orderly.network/ui-tradingview@2.8.1-alpha.0
+  - @orderly.network/ui-transfer@2.8.1-alpha.0
+  - @orderly.network/utils@2.8.1-alpha.0
+
 ## 2.8.0
 
 ### Minor Changes