+++ packages/ui-share/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/ui-share
 
+## 2.8.1
+
+### Patch Changes
+
+- publish
+- 0474e5f: publish
+- Updated dependencies
+- Updated dependencies [0474e5f]
+  - @orderly.network/hooks@2.8.1
+  - @orderly.network/i18n@2.8.1
+  - @orderly.network/types@2.8.1
+  - @orderly.network/ui@2.8.1
+  - @orderly.network/utils@2.8.1
+
+## 2.8.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.8.1-alpha.0
+  - @orderly.network/i18n@2.8.1-alpha.0
+  - @orderly.network/types@2.8.1-alpha.0
+  - @orderly.network/ui@2.8.1-alpha.0
+  - @orderly.network/utils@2.8.1-alpha.0
+
 ## 2.8.0
 
 ### Minor Changes